refactor(server): extract code related to aws marketplace into its own module
