--- packages/amplication-server/src/core/auth/aws-marketplace/constant.ts
@@ -1,2 +1,0 @@
-export const AWS_MARKETPLACE_INTEGRATION_CALLBACK_PATH =
-  "/auth/aws-marketplace/callback";
