+++ packages/amplication-server/src/core/aws-marketplace/constant.ts
@@ -1,0 +1,5 @@
+export const AWS_MARKETPLACE_INTEGRATION_REQUEST_PATH =
+  "/integration/marketplace";
+
+export const AWS_MARKETPLACE_INTEGRATION_CALLBACK_PATH =
+  "/integration/marketplace/callback";
