+++ packages/amplication-server/src/core/aws-marketplace/registration-page.ts
